third_party/WebKit/LayoutTests/http/tests/credentialmanager/passwordcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: PasswordCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
        name: 'name',
        iconURL: 'https://example.com/icon.png'
    });

    verify_interface('PasswordCredential', credential, {
        id: 'string',
        name: 'string',
        iconURL: 'string',
        type: 'string'
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, 'https://example.com/icon.png');
    assert_equals(credential.type, 'password');

    assert_true(credential.toFormData() instanceof FormData);
}, 'Interfaces and attributes of PasswordCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new PasswordCredential({
            id: 'id',
            password: 'pencil',
            name: 'name',
            iconURL: '-'
        });
    });
}, 'Construct a PasswordCredential with an invalid icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
        name: 'name',
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.type, 'password');
    assert_true(credential.toFormData() instanceof FormData);

}, 'Construct a PasswordCredential with an empty icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, '');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.type, 'password');
    assert_true(credential.toFormData() instanceof FormData);
}, 'Construct a PasswordCredential with an empty name and icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
    });

    var fd = credential.toFormData();
    fd.append('n1', 'v1');
    fd.append('n2', 'v2');
    fd.append('n3', 'v3');
    fd.append('n1', 'v4');
    fd.append('n2', 'v5');
    fd.append('n3', 'v6');

    assert_equals(fd.get('n1'), null);
    assert_equals(fd.get('n2'), null);
    assert_equals(fd.get('n3'), null);

    assert_equals(fd.has('n1'), false);
    assert_equals(fd.has('n2'), false);
    assert_equals(fd.has('n3'), false);

    assert_array_equals(fd.getAll('n1'), []);
    assert_array_equals(fd.getAll('n2'), []);
    assert_array_equals(fd.getAll('n3'), []);

    for(var entry of fd) {
        assert_unreached("There should be nothing to iterate here.");
    }
}, 'Verify properties of opaque FormData.');

function verifyBodyMethodBehavior(t, req, methodName) {
    req[methodName]().then(
        t.unreached_func("The 'Body::" + methodName + "()' method should reject."),
        t.step_func_done(function (e) { assert_equals(e.name, "TypeError"); }));
}

// TODO(mkwst): 'Body' doesn't yet implement 'formData()'.
['arrayBuffer', 'blob', 'json', 'text'].forEach(function (methodName) {
    async_test(function (t) {
        var credential = new PasswordCredential({
            id: 'id',
            password: 'pencil',
        });
        var fd = credential.toFormData();
        fd.append('n1', 'v1');
        fd.append('n2', 'v2');
        fd.append('n3', 'v3');
        fd.append('n1', 'v4');
        fd.append('n2', 'v5');
        fd.append('n3', 'v6');

        t.step(function () {
            var r = new Request("http://127.0.0.1:8000", { body: fd, method: "POST" });
            verifyBodyMethodBehavior(t, r, methodName);
        });

    }, "Verify behavior of 'Body::" + methodName + "()' for opaque Requests.");
});
</script>
